Add a bunch more tests to the test suite, some of which require a 32-bit system.
These tests, which I mostly wrote a while ago but am just now committing, give
me some confidence that the conversion code is actually correct!  I couldn't
find a practical word-size-independent way to test such things as
stringToBigUnsigned("4294967296").toUnsignedLong() , so for now I'm making the
test suite specific to systems with short = 16 bits and int = long = 32 bits and
hoping that, if it works on those, it will also work on others.